refactor(p2p): split message processing from libp2p networking
Extract received-message handling (validation, persistence, consensus callbacks, and the reqresp data handlers) out of LibP2PService into a new P2PMessageProcessor. The processor is built in the factory (the composition root) and injected into both the libp2p service and the P2PClient; consensus callbacks now register directly on the processor rather than on the networking service. LibP2PService is reduced to the networking layer.

No behaviour change. This draws the boundary along the already-mostly-async P2PService interface in preparation for moving the libp2p stack onto a worker thread.
